% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data2catalog.R
\name{data2catalog}
\alias{data2catalog}
\title{Convert parsed data to catalog}
\usage{
data2catalog(X)
}
\arguments{
\item{X}{parsed data}
}
\value{
list of: \code{catalog} - a catalog;  and \code{root} - character id of root node.
}
\description{
Turn parsed data around a tree and returns it in in catalog form
}
\details{
The input data format is that given as output from \code{\link{parseData}} and used as input by \code{\link{makeTree}}. A catalog (output) is a list of nodes with data on their associated cost, QoL, children and probabilities for each child. The parsed data (input) is a list of nodes (their costs/QoLs), edges and probabilities. The identity of the root is also returned as needed by \code{\link{makeTree}} in building the tree.
}
\author{
Pete Dodd
}

