% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dtree.R
\name{dtree}
\alias{dtree}
\title{Main user function for \code{dtree} package}
\usage{
dtree(ss, view = TRUE)
}
\arguments{
\item{ss}{input string defining model}

\item{view}{logical for returning data or viewing graph}
}
\value{
list of \itemize{
  \item \code{coststring} - string giving mean cost for model
  \item \code{qolstring} - string giving mean QoL for model
  \item \code{costfun} - function for calculating cost for model
  \item \code{qolfun} - function for calculating QoL for model
  \item \code{root} - string identifying the root of model
  \item \code{dot} - string in DOT language for export and visualization
  \item \code{inputstring} - the string used as input.
}
}
\description{
Main user function for \code{dtree} package, wrapping most other functions
}
\details{
This function takes a string in the syntax described in the vignette (type \code{vignette('dtree')} to view) and returns various related data. If \code{view=TRUE}, there is not return, but a browser is opened to view the decision tree specified via the \code{DiagrammeR} package. If \code{view=FALSE}, strings defining the mean costs and QoLs for the decision tree are returned, as well as functions which apply these to PSA data in the form of \code{data.frame}s. A DOT string is also returned for export and generation of graphs via Graphviz, as is the original input string.
}
\author{
Pete Dodd
}

