% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getStings.R
\name{getCstring}
\alias{getCstring}
\title{Make cost string}
\usage{
getCstring(x)
}
\arguments{
\item{x}{a tree}
}
\value{
string
}
\description{
Obtain a cost string from a tree
}
\details{
This is intended to be used on tree output from \code{\link{makeTree}} and recursively traverse through to generate a cost string, for likely use in \code{\link{string2fun}}.
}
\author{
Pete Dodd
}

