% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/makeTree.R
\name{makeTree}
\alias{makeTree}
\title{Make a tree}
\usage{
makeTree(nm, x)
}
\arguments{
\item{nm}{a character specifying the tree root}

\item{x}{tree data in catalog form}
}
\value{
tree
}
\description{
Turn tree data in catalog form into a tree data structure
}
\details{
This would normally be applied to output from \code{\link{data2catalog}} (which determines the identity of root and reshapes data into catalog form). Catalog form (input) is node-by-node data with information on children and their probabilities. The tree structure returned is the same data in a recursive list, for onward use in the functions \code{\link{getQstring}} and \code{\link{getCstring}}.
}
\author{
Pete Dodd
}

