% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{stability}
\alias{stability}
\title{Main function to calculate stability coefficients}
\usage{
stability(formula, data, methods = c("lm", "rpart", "tree", "ctree",
  "evtree"), n.rep = 100, weights, perc.sub = 0.75, prune = TRUE)
}
\arguments{
\item{formula}{a formula, weight a response to left of ~.}

\item{data}{Data frame to run models on}

\item{methods}{Which tree methods to use. Defaults:
lm, rpart, tree, ctree, evtree. Also can use "rf" for random forests}

\item{n.rep}{Number of times to replicate each method}

\item{weights}{Optional weights for each case.}

\item{perc.sub}{What fraction of data to put into train dataset. 1-frac.sub
is allocated to test dataset. Defaults to 0.75}

\item{prune}{Whether to prune rpart tree}
}
\description{
Main function to calculate stability coefficients
}
