% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/string2fun.R
\name{string2fun}
\alias{string2fun}
\title{String to function}
\usage{
string2fun(ss)
}
\arguments{
\item{ss}{a string}
}
\value{
a function
}
\description{
This function is designed to be called on a string and retruns a function (via closure) that evaluates the input string.
}
\details{


The input string is most likely to be generated as output from \code{\link{getCstring}} or \code{\link{getQstring}}. A list of inputs is evaluated to a single number using the eval command. N.B. this means that the variables in the string must be present, and that the input string must be a valid (valuable) arithmetic expression. If a data.frame is supplied, this is detected, and a vector of outputs returned with each a value for each row in the input data.frame.
}
\author{
Pete Dodd
}

