% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.dtree}
\alias{summary.dtree}
\title{Summary results from dtree.}
\usage{
summary.dtree(object, ...)
}
\arguments{
\item{object}{An object from dtree.}

\item{...}{Other arguments.}
}
\description{
Summary results from dtree.
}

