% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wrap4diag.R
\name{wrap4diag}
\alias{wrap4diag}
\title{Convert model definition string as DOT language}
\usage{
wrap4diag(x)
}
\arguments{
\item{x}{model definition string}
}
\value{
string in DOT language defining tree
}
\description{
Convert model definition string as DOT language
}
\details{
This is a helper function that takes the string defining the decision tree, strips out comments etc. and wraps it to make a valid DOT language string for use by \code{DiagrammeR} for visualisation, or export for use in Graphviz etc.
}
\author{
Pete Dodd
}

